+++ src/NS.h
@@ -40,10 +40,10 @@ public:
     // member functions - alphabetically
     static  DBL_DBL_DBL CalculateCoreCollapseSNParams_Static(const double p_Mass);
 
-    static  DBL_DBL_DBL DeltaAngularMomentumByPulsarAccretion(const double p_MassGainPerTimeStep, 
-                                                const double p_MagField, 
+    static  DBL_DBL_DBL DeltaAngularMomentumByPulsarAccretion_Static(const double p_MassGainPerTimeStep, 
                                                 const double p_Mass, 
                                                 const double p_Radius,
+                                                const double p_MagField, 
                                                 const double p_SpinFrequency, 
                                                 const double p_AngularMomentum, 
                                                 const double p_Stepsize,