+++ src/typedefs.h
@@ -701,7 +701,7 @@ const COMPASUnorderedMap<NEUTRINO_MASS_LOSS_PRESCRIPTION, std::string> NEUTRINO_
 
 // neutron star accretion scenario under common envelope
 enum class NS_ACCRETION_IN_CE: int { ZERO, SURFACE, DISK };
-const COMPASUnorderedMap<NS_ACCRETION_IN_CE, std::string> NS_ACCRETION_IN_CELabel = {
+const COMPASUnorderedMap<NS_ACCRETION_IN_CE, std::string> NS_ACCRETION_IN_CE_LABEL = {
     { NS_ACCRETION_IN_CE::ZERO,  "ZERO" },
     { NS_ACCRETION_IN_CE::SURFACE, "SURFACE" },
     { NS_ACCRETION_IN_CE::DISK, "DISK" },
@@ -709,7 +709,7 @@ const COMPASUnorderedMap<NS_ACCRETION_IN_CE, std::string> NS_ACCRETION_IN_CELabe
 
 // neutron star equations of state
 enum class NS_EOS: int { SSE, ARP3 };
-const COMPASUnorderedMap<NS_EOS, std::string> NS_EOSLabel = {
+const COMPASUnorderedMap<NS_EOS, std::string> NS_EOS_LABEL = {
     { NS_EOS::SSE,  "SSE" },
     { NS_EOS::ARP3, "ARP3" }
 };